+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.2
+
+- `Keyframe.at` static method to create empty `Keyframe` at offset
+- `Keyframes.create`, `Keyframes.createAt`, `Keyframes.createAppend`, `Keyframes.createAppendAt` methods
+to create and add/append keyframes directly
+
 ## 0.0.1
 
-- Initial version, created by Stagehand
+- `AnimationBuilder`, `Keyframe`, `Keyframes`, `AnimationOptions`